[libs] E: Extract sandbox config types
The configuration structures SandboxCacheConfig, SimpleSandboxCacheConfig,
and StandaloneConfig were previously defined inside the crates that also
carry heavy runtime machinery (nanvix-sandbox-cache, nanvix-sandbox, and
nanvix-http respectively). This created unnecessary coupling: consumers
that only need the config types had to depend on the full runtime crates.

Introduce a new lightweight nanvix-sandbox-config crate that owns all
three config structures, with minimal dependencies (hwloc, syscomm, and
optionally linuxd). The original locations now re-export from the new
crate so downstream paths through nanvix::sandbox_config remain stable.

Key changes:

- Create src/libs/nanvix-sandbox-config with modules multi_process,
  single_process (behind the single-process feature), and standalone.

- Move SandboxCacheConfig from nanvix-sandbox-cache::config and
  SimpleSandboxCacheConfig from nanvix-sandbox::simple_cache::config
  into the new crate. The old modules become thin re-export shims.

- Move StandaloneConfig from nanvix-http::client::standalone into the
  new crate. Make its fields private and add getter methods
  (kernel_binary_path, ramfs_filename, console_file, snapshot_path,
  mount_directory, gdb_port) for consistency with the accessor pattern
  used by SandboxCacheConfig and SimpleSandboxCacheConfig.

- Thread the nanvix-sandbox-config dependency and its feature flags
  (single-process, gdb) through nanvix-http, nanvix-sandbox,
  nanvix-sandbox-cache, nanvix-terminal, and nanvix.

- Re-export the crate from nanvix::sandbox_config so nanvixd and other
  top-level consumers continue to work without path changes.

- Update Makefile ALL_HOST_RUST_LIBS and workspace Cargo.toml members.

Validated: builds pass for microvm with multi-process, single-process,
and standalone deployment modes.
