+++ src/libs/nanvix-sandbox-config/Cargo.toml
@@ -1,0 +1,24 @@
+# Copyright(c) The Maintainers of Nanvix.
+# Licensed under the MIT License.
+
+[package]
+name = "nanvix-sandbox-config"
+version.workspace = true
+license-file.workspace = true
+edition.workspace = true
+authors.workspace = true
+description = "Sandbox configuration structures for Nanvix"
+homepage.workspace = true
+
+[dependencies]
+hwloc = { workspace = true }
+linuxd = { workspace = true, optional = true }
+syscomm = { workspace = true }
+
+[features]
+default = []
+single-process = ["linuxd"]
+gdb = []
+
+[lints]
+workspace = true
