+++ src/libs/nanvix-terminal/Cargo.toml
@@ -17,14 +17,24 @@ libc = { workspace = true }
 log = { workspace = true }
 nanvix-sandbox = { workspace = true, optional = true }
 nanvix-sandbox-cache = { workspace = true, optional = true }
+nanvix-sandbox-config = { workspace = true, optional = true }
 tokio = { workspace = true, features = ["full"] }
 user-vm-api = { workspace = true }
 uservm = { workspace = true, optional = true }
 
 [features]
 default = ["microvm", "multi-process"]
-single-process = ["nanvix-sandbox", "nanvix-sandbox/single-process"]
-multi-process = ["nanvix-sandbox-cache", "nanvix-sandbox-cache/multi-process"]
+single-process = [
+    "nanvix-sandbox-config",
+    "nanvix-sandbox-config/single-process",
+    "nanvix-sandbox",
+    "nanvix-sandbox/single-process",
+]
+multi-process = [
+    "nanvix-sandbox-config",
+    "nanvix-sandbox-cache",
+    "nanvix-sandbox-cache/multi-process",
+]
 standalone = ["config", "uservm"]
 gdb = ["uservm?/gdb"]
 # Enables fine-grained time profiling.
