+++ src/libs/nanvix/src/lib.rs
@@ -93,6 +93,7 @@ pub use nanvix_sandbox as sandbox;
 pub use nanvix_sandbox::simple_cache as sandbox_cache;
 #[cfg(feature = "multi-process")]
 pub use nanvix_sandbox_cache as sandbox_cache;
+pub use nanvix_sandbox_config as sandbox_config;
 pub use nanvix_terminal as terminal;
 pub use syscomm;
 pub use syslog as log;
